modules/module_A/requirements.txt
fastapi==0.115.2
uvicorn==0.29.0
pydantic==2.6.1

openai-whisper==20231117
torch==2.2.0
numpy==1.26.4

python-multipart==0.0.6
ffmpeg-python==0.2.0
